MS2024018/requirements.txt
bert-score==0.3.13
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.0
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.3.2
huggingface-hub==0.31.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy<2.0
pandas==2.2.3
Pillow==11.2.1
progressbar2==4.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.3.0
six==1.17.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
urllib3==2.4.0
zipp==3.21.0

# LoRA / PEFT
peft==0.5.0
accelerate==0.22.0
sentencepiece==0.1.99


